I ferico template

See merge request go/templatev2!2

+++ internal/usecases/task/impl/delete.go
@@ -6,7 +6,7 @@ import (
 	"go.elastic.co/apm"
 )
 
-func (u *usecase) DeleteTask(ctx context.Context, id int) (err error) {
+func (u *usecase) DeleteTask(ctx context.Context, id int64) (err error) {
 	span, _ := apm.StartSpan(ctx, "usecase", "DeleteTask")
 	defer span.End()
